inferencia/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
opencv-python-headless==4.8.1.78
ultralytics==8.0.196
aiohttp==3.9.1
pyyaml==6.0.1
numpy==1.24.3
pillow==10.1.0
torch==2.0.1
torchvision==0.15.2

